requirements.txt
asyncio
aiohttp
psycopg2-binary
redis